requirements.txt
flask
flask_limiter
requests
dnspython
beautifulsoup4
fake-useragent
python-dotenv